pyproject.toml
[project]
name = "phys-doc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymupdf>=1.26.3",
    "pymupdf4llm>=0.0.26",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]
